# docs/openapi.json
{
  "components": {
    "responses": {
...
      "BadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Validation failed."
      },
...
      "Forbidden": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "The key lacks the required scope."
      },
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "No such resource in this tenant."
      },
...
      "Unauthorized": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Missing or unknown bearer key."
      }
    },
    "schemas": {
...
      "BeliefReadResponse": {
        "additionalProperties": false,
        "properties": {
          "beliefId": {
            "type": "string"
          },
          "confidence": {
            "type": "number"
          },
          "conversationCount": {
            "type": "number"
          },
          "conversationIds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "corroborationCount": {
            "type": "number"
          },
          "field": {
            "type": "string"
          },
          "priorValue": {
            "type": "string"
          },
          "promoterVersion": {
            "type": "string"
          },
          "revision": {
            "type": "number"
          },
          "sourceSceneIds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "statement": {
            "type": "string"
          },
          "statementSource": {
            "enum": [
              "template",
              "llm"
            ],
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "supersededBy": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "validFrom": {
            "type": "string"
          },
          "validUntil": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "beliefId",
          "userId",
          "subject",
          "field",
          "value",
          "statement",
          "statementSource",
          "confidence",
          "revision",
          "status",
          "validFrom",
          "sourceSceneIds",
          "conversationIds",
          "corroborationCount",
          "conversationCount"
        ],
        "type": "object"
      },
      "BeliefsListResponse": {
        "additionalProperties": false,
        "properties": {
          "beliefs": {
            "items": {
              "$ref": "#/components/schemas/BeliefReadResponse"
            },
            "type": "array"
          },
          "found": {
            "type": "number"
          }
        },
        "required": [
          "beliefs",
          "found"
        ],
        "type": "object"
      },
...
      "IngestMentionRequest": {
        "additionalProperties": false,
        "properties": {
          "contextRef": {
            "$ref": "#/components/schemas/MentionContextRef"
          },
          "emittedAt": {
            "type": "string"
          },
          "knownEntities": {
            "items": {
              "$ref": "#/components/schemas/KnownEntity"
            },
            "type": "array"
          },
          "text": {
            "maxLength": 16000,
            "type": "string"
          },
          "timezone": {
            "maxLength": 64,
            "type": "string"
          },
          "userId": {
            "maxLength": 200,
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "API key (`Authorization: Bearer <key>`); its SHA-256 must be registered in BRAIN_API_KEYS. Keys carry scopes (`brain:read`, `brain:write`, `brain:admin`, `brain:read_pii`, `registry:publish`, `registry:curate`, `indexer:write`) — each operation states the scope it requires. Not an OAuth flow.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Generated from the zod wire contracts (`pnpm openapi:build` — scripts/build-openapi.ts); do not edit by hand. Excluded on purpose: the operator/ops surface (jobs, leases, policy, stats, maintenance, GDPR cascades), which is indexed in docs/api.md, and the MCP transport `ALL /mcp/{companyId}`, which is JSON-RPC over Streamable HTTP rather than REST — MCP clients discover it through `tools/list` (docs/mcp.md). Scopes are plain bearer-key grants, not OAuth flows — each operation states its required scope.",
    "license": {
      "identifier": "AGPL-3.0-or-later",
      "name": "AGPL-3.0-or-later"
    },
    "summary": "The community-facing platform surface: the memory core (fact ingest, search, multi-hop, synthesize, entity and fact reads), the Domain Pack registry and tenant pack admin, the document pipeline, the external-indexer work protocol, the raw-substrate driver, the evidence substrate, and source reputation reads.",
    "title": "INITE Brain — Platform API",
    "version": "2.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/beliefs": {
      "get": {
        "description": "Beliefs the brain currently holds (semantic_belief, promoted from enriched scenes), filtered by exact subject/field key, lifecycle status (default active) and user scope. A user-bound token is pinned to its own user (userId mismatch is 403); M2M keys may scope to any user or list tenant-wide. Page capped at 100 (default 25). On by default; BELIEFS_API_ENABLED=0 makes it a 404. Source: src/beliefs/beliefs.controller.ts.\n\nRequired scope: `brain:read`.",
        "operationId": "listBeliefs",
        "parameters": [
          {
            "description": "Free-text subject key (exact match).",
            "in": "query",
            "name": "subject",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Free-text field key (exact match).",
            "in": "query",
            "name": "field",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "End-user scope key. M2M keys may assert any user; user-bound tokens are pinned to their own.",
            "in": "query",
            "name": "userId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Lifecycle filter: `active` (default), `superseded`, `all`.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Page size (default 25, max 100).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeliefsListResponse"
                }
              }
            },
            "description": "The visible beliefs."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "List beliefs by their free-text (subject, field) key",
        "tags": [
          "Beliefs"
        ]
      }
    },
    "/v1/beliefs/{id}": {
      "get": {
        "description": "One semantic_belief revision as stored: the held value, its rendered statement, confidence, the supersede chain (revision/status/supersededBy/validFrom/validUntil), inline scene provenance (sourceSceneIds) and corroboration counters. Superseded revisions still resolve. Every visibility fence (tenant, fail-closed single-user scope) answers 404 — existence never leaks. On by default; BELIEFS_API_ENABLED=0 makes it a 404. Source: src/beliefs/beliefs.controller.ts.\n\nRequired scope: `brain:read`.",
        "operationId": "getBelief",
        "parameters": [
          {
            "description": "Belief record id (`semantic_belief:…`).",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeliefReadResponse"
                }
              }
            },
            "description": "The belief."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Read one belief revision by id",
        "tags": [
          "Beliefs"
        ]
      }
    },
...
    "/v1/entities/{id}": {
      "get": {
        "description": "The entity plus the facts currently held about it, filtered by the row policy (a fact whose predicate requires a scope the caller lacks never appears). `asOf` slices WORLD time (what was true at T); `recordedAt` slices TRANSACTION time (what the graph believed at T — a later retract or supersede is ignored). A merged entity answers with `mergedInto` set: treat it as a redirect.\n\nRequired scope: `brain:read`.",
        "operationId": "getEntityProfile",
        "parameters": [
          {
            "description": "Entity id — short (`cuid_abc`) or full (`knowledge_entity:cuid_abc`).",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "ISO-8601 world-time slice.",
            "in": "query",
            "name": "asOf",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "ISO-8601 transaction-time slice.",
            "in": "query",
            "name": "recordedAt",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityProfileResponse"
                }
              }
            },
            "description": "The profile."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Entity profile and its active facts",
        "tags": [
          "Entities"
        ]
      }
    },
...
    "/v1/facts/{id}": {
      "get": {
        "description": "The fact itself — what is remembered, where it came from (vertical/recorder/conversation), its validity instant and user scope. Every visibility fence (tenant, user scope, row policy) answers 404 — existence never leaks. 404 until FACTS_API_ENABLED=1. Source: src/facts/facts.controller.ts.\n\nRequired scope: `brain:read`.",
        "operationId": "getFact",
        "parameters": [
          {
            "description": "Fact record id (`knowledge_fact:…`).",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactReadResponse"
                }
              }
            },
            "description": "The fact."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Read one fact by id",
        "tags": [
          "Facts"
        ]
      }
    },
...
    "/v1/ingest/fact": {
      "post": {
        "description": "The headline write. Resolves the entity, embeds the claim and runs bitemporal conflict resolution against the standing timeline: the response `outcome` says what the resolver decided (INSERTED / INSERTED_HISTORICAL / CORROBORATED / SUPERSEDED / COMPETING / REJECTED), never just \"ok\". `userId` stamps a per-user memory scope (migration 0055) — scope-local conflict resolution, invisible to every other user of the tenant and to requests that don't assert one. `explain: true` adds the deterministic conflict narrative on a SUPERSEDED / COMPETING outcome. Source: src/ingest/ingest.controller.ts.\n\nRequired scope: `brain:write`.",
        "operationId": "ingestFact",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestFactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestFactResponse"
                }
              }
            },
            "description": "The resolver's decision."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "summary": "Record a declared structured fact",
        "tags": [
          "Ingest"
        ]
      }
    },
    "/v1/ingest/mention": {
      "post": {
        "description": "The simplest write, and the one to reach for when the caller has prose rather than a structured claim: a chat turn, a meeting note, an email body. The text is captured as an episode, an LLM extracts entities, facts and edges from it, and each extracted fact goes through the same bitemporal resolver as `POST /v1/ingest/fact` — so extraction is noisy by design and conflict resolution cleans up at read time. Only `text` is required: `contextRef` defaults to the `chat` vertical and `emittedAt` to the moment the request arrives. `userId` stamps the per-user memory scope on the episode and every extracted fact. `skipped: true` means the extractor found nothing worth recording (`reason` says which check stopped it) — that is a normal outcome, not an error. Prefer `POST /v1/ingest/document` for anything long enough to need chunking. Source: src/ingest/ingest.controller.ts.\n\nRequired scope: `brain:write`.",
        "operationId": "ingestMention",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestMentionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestMentionResponse"
                }
              }
            },
            "description": "What the extractor wrote."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "summary": "Record free text and let brain extract the facts",
        "tags": [
          "Ingest"
        ]
      }
    },
    "/v1/keys": {
      "get": {
        "description": "The tenant's `companyId`, its MCP URL, and the keys visible to the caller — a user-bound credential sees the keys bound to that user, a `brain:admin` (or M2M) credential sees every key in the tenant. Secrets and hashes are never returned: a row carries a display `prefix` only. `issuableScopes` is what THIS credential may grant, and `issuingEnabled` is false on a deployment with no database-backed key store. Source: src/keys/keys.controller.ts.\n\nRequired scope: `brain:read`.",
        "operationId": "listApiKeys",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyListResponse"
                }
              }
            },
            "description": "Visible keys plus the connection values."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "summary": "List the API keys this credential can see",
        "tags": [
          "Keys"
        ]
      },
      "post": {
        "description": "Mints a key brain verifies itself, returning the plaintext ONCE — it is stored only as a SHA-256 hash and cannot be recovered. The granted scopes are the requested ones intersected with what the calling credential already holds, so a key is never wider than its issuer; asking for none that the caller holds is 403. Only the four delegable scopes are accepted (`brain:read`, `brain:write`, `brain:admin`, `brain:read_pii`) — operator scopes are not mintable. A user-bound caller can only bind the key to itself. Tenants are capped at 25 live keys.\n\nRequired scope: `brain:read`.",
        "operationId": "issueApiKey",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedKeyResponse"
                }
              }
            },
            "description": "The key, shown once."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "summary": "Issue an API key",
        "tags": [
          "Keys"
        ]
      }
    },
...
    "/v1/scenes": {
      "get": {
        "description": "Scenes of the current segmenter world (memory_episode: coherent multi-turn groupings the composer derives from raw turns), newest first, with their member episodes, knowledge-graph backlinks, enrichment-owned state deltas (the beliefs are promoted from these) and per-dimension memory value. Filter by conversationId, entityId, since/until (overlap) and user scope. A user-bound token is pinned to its own user (userId mismatch is 403) and sees its own scenes plus tenant-global ones whose member set contains it; an M2M key scopes to any user with userId, and without one lists tenant-global scenes only (gists quote verbatim turns — the episodes contract). Scenes carrying PII need brain:read_pii. Page capped at 100 (default 25). `world` names the version served (empty until a world has been composed). On by default; SCENES_API_ENABLED=0 makes it a 404. Source: src/scenes/scenes.controller.ts.\n\nRequired scope: `brain:read`.",
        "operationId": "listScenes",
        "parameters": [
          {
            "description": "Only scenes covering this conversation.",
            "in": "query",
            "name": "conversationId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only scenes back-linked to this entity (`knowledge_entity:…`).",
            "in": "query",
            "name": "entityId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "End-user scope key. M2M keys may assert any user; user-bound tokens are pinned to their own.",
            "in": "query",
            "name": "userId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "ISO date-time: scenes ending at or after this instant.",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "ISO date-time: scenes starting at or before this instant.",
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Page size (default 25, max 100).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenesListResponse"
                }
              }
            },
            "description": "The visible scenes."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "List scenes — what happened, together, when",
        "tags": [
          "Scenes"
        ]
      }
    },
    "/v1/scenes/{id}": {
      "get": {
        "description": "One memory_episode scene as stored — the record a belief's sourceSceneIds and a scene citation point at. Resolves in any world (a superseded world's scene still opens). Every visibility fence (tenant, user scope, PII) answers 404 — existence never leaks. On by default; SCENES_API_ENABLED=0 makes it a 404. Source: src/scenes/scenes.controller.ts.\n\nRequired scope: `brain:read`.",
        "operationId": "getScene",
        "parameters": [
          {
            "description": "Scene record id (`memory_episode:…`).",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneReadResponse"
                }
              }
            },
            "description": "The scene."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "Read one scene by id",
        "tags": [
          "Scenes"
        ]
      }
    },
    "/v1/search": {
      "post": {
        "description": "The headline read. Hybrid retrieval by default (vector + BM25 fused by reciprocal rank), router-boosted, listwise-reranked, then backfilled with each hit entity’s facts. Default semantics is Datomic-style \"actual now\" — only facts whose validity window contains the query moment; `asOf` slices world-time and `includeStale` reverts to the audit shape. `userId` widens the fence from tenant-global memory to tenant-global PLUS that user's personal rows (fail-closed: omitted means global only). Facts whose predicate requires a scope the caller lacks never enter the result. Fans out to the shared embedding / LLM budget, so it sits in the per-credential `expensive` throttle bucket — 10 requests/min, not the 120/min default. Source: src/search/search.controller.ts.\n\nRequired scope: `brain:read`.",
        "operationId": "searchKnowledge",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Scored entity hits with their facts."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "summary": "Search the knowledge graph",
        "tags": [
          "Search"
        ]
      }
    },
...
    "/v1/synthesize": {
      "post": {
        "description": "Corrective RAG over the same retrieval stack as /v1/search: generate an answer, then verify each claim against the facts it cites. The pipeline ABSTAINS rather than guess — `answer` is null and `reason` names the gate that fired (no results, thin coverage, ungrounded support, a failed verifier…). `synthesisGuardrails` picks the policy: `strict` closes to null on a partial verdict, `lenient` returns the answer with the verdict attached, `off` skips the verifier. Cited facts are always the caller’s own visible memory. Fans out to the shared embedding / LLM budget, so it sits in the per-credential `expensive` throttle bucket — 10 requests/min, not the 120/min default. Source: src/synthesize/synthesize.controller.ts.\n\nRequired scope: `brain:read`.",
        "operationId": "synthesizeAnswer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SynthesizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SynthesizeResponse"
                }
              }
            },
            "description": "The answer (or a reasoned abstention)."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "summary": "Answer a question from memory, with citations",
        "tags": [
          "Search"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Hosted instance (placeholder — self-hosted deployments serve the same paths on their own origin).",
      "url": "https://brain.inite.ai"
    }
  ],
...
}